src/requirements.txt
aiogram==3.10.0
motor==3.4.0
pymongo==4.7.3
python-decouple==3.8